BlackoilModel: add impl file
